refactor(ddd): improve folder structure to avoid mixing adapter pattern with ddd
