--- .vscode/launch.json
@@ -1,19 +1,0 @@
-{
-  "version": "0.2.0",
-  "configurations": [
-    {
-      "type": "node",
-      "request": "launch",
-      "name": "Debug Nest Framework",
-      "runtimeExecutable": "npm",
-      "runtimeArgs": ["run", "start:debug", "--", "--inspect-brk"],
-      "autoAttachChildProcesses": true,
-      "restart": true,
-      "sourceMaps": true,
-      "stopOnEntry": false,
-      "console": "integratedTerminal",
-      "rootPath": "${workspaceFolder}",
-      "cwd":"${workspaceFolder}"
-    }
-  ]
-}
