+++ src/modules/prisma/prisma.module.ts
@@ -1,5 +1,5 @@
 import { Module } from '@nestjs/common';
-import { PrismaService } from './prisma.service';
+import { PrismaService } from './app/service/prisma.service';
 
 @Module({
   providers: [PrismaService],
