+++ src/modules/prometheus/prometheus.module.ts
@@ -1,6 +1,6 @@
 import { Module } from '@nestjs/common';
 import { PrometheusService } from '@src/modules/prometheus/app/service/prometheus.service';
-import { PrometheusController } from '@src/modules/prometheus/ports/prometheus.controller';
+import { PrometheusController } from '@src/modules/prometheus/interfaces/prometheus.controller';
 
 @Module({
   providers: [PrometheusService],
