+++ src/modules/recipe/app/command/create-recipe.command.ts
@@ -1,5 +1,5 @@
 import { ICommand } from '@nestjs/cqrs';
-import { CreateRecipeDto } from '../../domain/recipe/create-recipe.dto';
+import { CreateRecipeDto } from '@src/modules/recipe/app/dtos/create-recipe.dto';
 
 export class RecipeCreateCommand implements ICommand {
   constructor(public readonly data: CreateRecipeDto) {}
