+++ src/modules/recipe/app/command/update-recipe.command.ts
@@ -1,5 +1,5 @@
 import { ICommand } from '@nestjs/cqrs';
-import { UpdateRecipeDto } from '@src/modules/recipe/domain/recipe/update-recipe.dto';
+import { UpdateRecipeDto } from '@src/modules/recipe/app/dtos/update-recipe.dto';
 
 export class UpdateRecipeCommand implements ICommand {
   constructor(
