+++ src/modules/recipe/app/query/find-recipe.handler.spec.ts
@@ -1,13 +1,13 @@
 import { Test, TestingModule } from '@nestjs/testing';
-import { RecipeFactory } from '@src/modules/recipe/factories/recipe.factory';
+import { RecipeFactory } from '@src/modules/recipe/test/factories/recipe.factory';
 import { PrismaModule } from '@src/modules/prisma/prisma.module';
 import { CqrsModule } from '@nestjs/cqrs';
-import { Repository } from '../../entities/recipe.repository';
-import { RecipeRepository } from '../../adapters/recipe.repository';
+import { Repository } from '@src/modules/recipe/domain/recipe/repositories/recipe.repository.interface';
+import { RecipeRepository } from '@src/modules/recipe/infrastructure/repositories/recipe.repository';
 import { FindRecipeHandler } from '@src/modules/recipe/app/query/find-recipe.handler';
 import { FindRecipeQuery } from '@src/modules/recipe/app/query/find-recipe.query';
 import { faker } from '@faker-js/faker';
-import { RecipeNotFoundException } from '@src/modules/recipe/app/exception/recipe-not-found.exception';
+import { RecipeNotFoundException } from '@src/modules/recipe/domain/exception/recipe-not-found.exception';
 
 const recipeRepositoryMock: jest.Mocked<Repository> = {
   create: jest.fn(),
