+++ src/modules/recipe/app/query/find-recipe.handler.ts
@@ -1,8 +1,8 @@
 import { Injectable } from '@nestjs/common';
 import { IQueryHandler, QueryHandler } from '@nestjs/cqrs';
 import { FindRecipeQuery } from '@src/modules/recipe/app/query/find-recipe.query';
-import { RecipeRepository } from '@src/modules/recipe/adapters/recipe.repository';
-import { RecipeNotFoundException } from '../exception/recipe-not-found.exception';
+import { RecipeRepository } from '@src/modules/recipe/infrastructure/repositories/recipe.repository';
+import { RecipeNotFoundException } from '@src/modules/recipe/domain/exception/recipe-not-found.exception';
 
 @Injectable()
 @QueryHandler(FindRecipeQuery)
