+++ src/modules/recipe/app/query/find-recipe.query.ts
@@ -1,5 +1,5 @@
 import { Query } from '@nestjs/cqrs';
-import { SingleRecipeDto } from '@src/modules/recipe/domain/recipe/single-recipe.dto';
+import { SingleRecipeDto } from '@src/modules/recipe/app/dtos/single-recipe.dto';
 
 export class FindRecipeQuery extends Query<SingleRecipeDto> {
   constructor(public readonly id: number) {
