+++ src/modules/recipe/domain/exception/recipe-not-found.exception.ts
@@ -1,4 +1,4 @@
-import { BaseException } from '@src/modules/common/exceptions/base.exception';
+import { BaseException } from '@src/modules/shared/exceptions/base.exception';
 
 export class RecipeNotFoundException extends BaseException {
   constructor(recipeId: number) {
