+++ src/modules/recipe/infrastructure/repositories/recipe.repository.int-spec.ts
@@ -4,9 +4,9 @@ import {
   PostgreSqlContainer,
   StartedPostgreSqlContainer,
 } from '@testcontainers/postgresql';
-import { PrismaService } from '@src/modules/prisma/prisma.service';
-import { RecipeRepository } from '@src/modules/recipe/adapters/recipe.repository';
-import { RecipeFactory } from '@src/modules/recipe/factories/recipe.factory';
+import { PrismaService } from '@src/modules/prisma/app/service/prisma.service';
+import { RecipeRepository } from '@src/modules/recipe/infrastructure/repositories/recipe.repository';
+import { RecipeFactory } from '@src/modules/recipe/test/factories/recipe.factory';
 import { RecipeCreateCommand } from '@src/modules/recipe/app/command/create-recipe.command';
 import { Logger } from '@nestjs/common';
 import { UpdateRecipeCommand } from '@src/modules/recipe/app/command/update-recipe.command';
