+++ src/modules/shared/exceptions/base.exception.ts
