+++ src/modules/shared/logging/logging.querybus.ts
